backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
asyncpg>=0.29.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
httpx>=0.26.0
openai>=1.12.0
anthropic>=0.18.1
google-generativeai>=0.3.2
langgraph>=0.0.26
langchain>=0.1.9
langchain-openai>=0.0.8
opentelemetry-api>=1.23.0
opentelemetry-sdk>=1.23.0
opentelemetry-instrumentation-fastapi>=0.44b0
opentelemetry-exporter-otlp>=1.23.0
prometheus-fastapi-instrumentator>=7.0.0
redis>=5.0.1
tenacity>=8.2.3
tiktoken>=0.6.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
python-dotenv>=1.0.1
pypdf>=4.0.0
pandas>=2.2.0
openpyxl>=3.1.2
tabulate>=0.9.0
